Ch. 1 Coles edits
# Things added
## Discussion of three patterns that may underlie a correlation

## Restructured and streamlined the prose

# Things removed

## Discussion of book mantra
Didn't have strong feelings about this, but felt that it interrupted the flow a little bit.

## Discussion of why experiments are common in psychology
Seemed unnecessary

## Scissor metaphor
Removed so that we could emphasize three reasons why two variables may be correlated

## Logic of comparison
Felt this was already nicely captured in a footnote

## Within vs. between subject designs
This seemed only tangentially related to the scope of the chapter--and I assume we'll cover it later.

## Removed ethics box
Replaced it with an in-text discussion of cross-cultural gratitude research

## Half dozen footnotes
This is idiosyncratic, but I strongly dislike footnotes. I think they're distracting for readers (particularly those who don't are ESOL and neurodivergent) and often unnecessary. Think about it: mid paragraph, I now need to avert my gaze to the footnote and then go back to the main text and try to figure out where I was and what the writing was talking about. My hot take is that the only people who actually appreciate footnotes are the people who write them. :P